rename /rag, update readme.md
